fix: resolve narrowing conversion bugs in file operations
- Fixed uninitialized variable 'bytesRead' in fileChecksum method
- Changed int to qint64 for file size operations to prevent overflow
- Fixed narrowing conversions in progress bar calculation using 100LL multiplier
- Added explicit cast for QMap::count() to int for XML tag counting
- All changes maintain Qt 5.15+ and Qt 6.x compatibility
- Prevents potential data loss with large files (>2GB) and calculation overflows

+++ apps/EKiosk/src/updater/textprogressbar.cpp
@@ -21,7 +21,7 @@ void TextProgressBar::update() {
     if (maximum > 0) {
         // we know the maximum
         // draw a progress bar
-        int percent = value * 100 / maximum;
+        int percent = static_cast<int>((value * 100LL) / maximum);
         int hashes = percent / 2;
 
         QByteArray progressbar(hashes, '#');
